# wrapper performance optimizations