--- a/ChangeLog
+++ b/ChangeLog
@@ -8,6 +8,11 @@ Darshan-3.3.1
   scenarios because of incomplete dependency information in Darshan shared
   library
 * modified Darshan to properly handle for non-MPI applications that call fork()
+* Performance optimizations to runtime wrappers to limit function call
+  overhead and locking costs
+* Replace mutex locks with atomics in core library
+* Add optional, limited support for RDTSCP-based timing, enabled via
+  --enable-rdtscp at configure time
 
 Darshan-3.3.0
 =============