--- a/darshan-runtime/lib/darshan-common.c
+++ b/darshan-runtime/lib/darshan-common.c
@@ -21,27 +21,6 @@
 
 #include "darshan.h"
 
-/* track opaque record referencre using a hash link */
-struct darshan_record_ref_tracker
-{
-    void *rec_ref_p;
-    UT_hash_handle hlink;
-};
-
-void *darshan_lookup_record_ref(void *hash_head, void *handle, size_t handle_sz)
-{
-    struct darshan_record_ref_tracker *ref_tracker;
-    struct darshan_record_ref_tracker *ref_tracker_head =
-        (struct darshan_record_ref_tracker *)hash_head;
-
-    /* search the hash table for the given handle */
-    HASH_FIND(hlink, ref_tracker_head, handle, handle_sz, ref_tracker);
-    if(ref_tracker)
-        return(ref_tracker->rec_ref_p);
-    else
-        return(NULL);
-}
-
 int darshan_add_record_ref(void **hash_head_p, void *handle, size_t handle_sz,
     void *rec_ref_p)
 {