--- a/darshan-runtime/lib/darshan-core-init-finalize.c
+++ b/darshan-runtime/lib/darshan-core-init-finalize.c
@@ -18,7 +18,6 @@
 #endif
 
 #include "darshan.h"
-#include "darshan-core.h"
 #include "darshan-dynamic.h"
 
 #ifdef HAVE_MPI