--- a/darshan-runtime/lib/darshan-core-init-finalize.c
+++ b/darshan-runtime/lib/darshan-core-init-finalize.c
@@ -30,6 +30,7 @@ int DARSHAN_DECL(MPI_Init)(int *argc, char ***argv)
     int ret;
 
     MAP_OR_FAIL(PMPI_Init);
+    (void)__darshan_disabled;
 
     ret = __real_PMPI_Init(argc, argv);
     if(ret != MPI_SUCCESS)
@@ -56,6 +57,7 @@ int DARSHAN_DECL(MPI_Init_thread)(int *argc, char ***argv, int required, int *pr
     int ret;
 
     MAP_OR_FAIL(PMPI_Init_thread);
+    (void)__darshan_disabled;
 
     ret = __real_PMPI_Init_thread(argc, argv, required, provided);
     if(ret != MPI_SUCCESS)
@@ -82,6 +84,7 @@ int DARSHAN_DECL(MPI_Finalize)(void)
     int ret;
 
     MAP_OR_FAIL(PMPI_Finalize);
+    (void)__darshan_disabled;
 
     darshan_core_shutdown(1);