--- a/darshan-runtime/lib/darshan-core.c
+++ b/darshan-runtime/lib/darshan-core.c
@@ -2451,6 +2451,7 @@ void darshan_core_fprintf(
     va_list ap;
 
     MAP_OR_FAIL(vfprintf);
+    (void)__darshan_disabled;
 
     va_start(ap, format);
     __real_vfprintf(stream, format, ap);