… darshan-runtime/doc/darshan-runtime.txt
… darshan-runtime/lib/darshan-core.h
… darshan-runtime/lib/darshan-hdf5.c
… darshan-runtime/lib/darshan-lustre.h
… darshan-runtime/lib/darshan-mdhim.c
… darshan-runtime/lib/darshan-mpiio.c
… darshan-runtime/lib/darshan-null.c
… darshan-runtime/lib/darshan-pnetcdf.c
… darshan-runtime/lib/darshan-posix.c
… darshan-runtime/lib/darshan-stdio.c
… darshan-runtime/lib/darshan.h
… darshan-test/regression/test-cases/src/stdio-benchmark.c