@carns 2021-07-09
Rebased this morning to a) consolidate commits more sensibly and eliminate reverted auxiliary timer library experiment and b) sync up with current origin/main

Will still need to rebase again after #381 .