@carns 2021-07-20
Rebased to current origin/main (i.e., to reflect automake changes)